add edges file
